+++ dimerreactioncapture_data.h
@@ -7,9 +7,9 @@
 class DimerReactionCaptureData : public ReactionData<DimerData>
 {
 public:
-    DimerReactionCaptureData(double k, int prevState, int nextState);
+    DimerReactionCaptureData(long double k, int prevState, int nextState);
 
-    double rate(const DimerData *dimer) const;
+    long double rate(const DimerData *dimer) const;
     void doIt(DimerData *const dimer) const;
 };
