+++ dimerreactionstabilization_data.cpp
@@ -1,9 +1,9 @@
 #include "dimerreactionstabilization_data.h"
 
-DimerReactionStabilizationData::DimerReactionStabilizationData(double k, int prevState, int nextState) :
+DimerReactionStabilizationData::DimerReactionStabilizationData(long double k, int prevState, int nextState) :
     ReactionData<DimerData>(k, prevState, nextState) {}
 
-double DimerReactionStabilizationData::rate(const DimerData *dimer) const {
+long double DimerReactionStabilizationData::rate(const DimerData *dimer) const {
     return (dimer->first->value() == prevState() && dimer->second->value() == prevState()) ?
                 rateValue(dimer) : 0;
 }
