+++ lotkathreereactor_context.h
@@ -9,7 +9,7 @@ public:
     LotkaThreeReactorContext();
 
     int numOfSpecs() const { return 3; }
-    double maxTime() const { return 5000; }
+    long double maxTime() const { return 5000; }
 
 protected:
     void solveToOut(std::ostream &out) const;
