+++ reactionactivation_data.cpp
@@ -1,9 +1,9 @@
 #include <cmath>
 #include "reactionactivation_data.h"
 
-ReactionActivationData::ReactionActivationData(const RTReactorContext *reactor, double Ea) :
+ReactionActivationData::ReactionActivationData(const RTReactorContext *reactor, long double Ea) :
     _reactor(reactor), _Ea(Ea) {}
 
-double ReactionActivationData::expFactor(double additionalEnergy) const {
+long double ReactionActivationData::expFactor(long double additionalEnergy) const {
     return exp(-(_Ea + additionalEnergy) / _reactor->RT());
 }
